Limpiar rama 16.0: Solo contenido específico de Odoo 16 (l10n_do_ncf)
