+++ __init__.py
@@ -1,2 +1,2 @@
 from . import models
-from . import wizards
+from . import reports
