--- l10n_do_ncf/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import reports
