--- l10n_do_ncf/__manifest__.py
@@ -1,22 +1,0 @@
-{
-    'name': 'Número de Comprobante Fiscal RD - RENACE',
-    'version': '1.0',
-    'category': 'Accounting',
-    'summary': 'Genera Número de Comprobante Fiscal para República Dominicana - Proyecto RENACE',
-    'author': 'Adderly Marte',
-    'website': 'https://renace.tech',
-    'license': 'LGPL-3',
-    'depends': ['account'],
-    'data': [
-        'security/ir.model.access.csv',
-        'views/ncf_fiscal_type_views.xml',
-        'views/account_move_views.xml',
-        'views/ncf_sequence_views.xml',
-        'views/res_config_settings_views.xml',
-        'reports/invoice_report.xml',
-        'data/fiscal_type_data.xml',
-    ],
-    'installable': True,
-    'application': False,
-    'auto_install': False,
-}
