--- l10n_do_ncf/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import account_move
-from . import ncf_fiscal_type
-from . import ncf_sequence
-from . import res_config_settings
