--- l10n_do_ncf/views/account_move_views.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_move_form_inherit_ncf" model="ir.ui.view">
-        <field name="name">account.move.form.inherit.ncf</field>
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account.view_move_form"/>
-        <field name="arch" type="xml">
-            <!-- Add intuitive buttons to header -->
-            <xpath expr="//header" position="inside">
-                <button name="toggle_use_ncf" 
-                        string="🧾 NCF" 
-                        type="object" 
-                        class="btn-outline-secondary"
-                        attrs="{'invisible': [('state', '!=', 'draft')]}"
-                        help="Activar o desactivar el uso de Número de Comprobante Fiscal"/>
-
-            </xpath>
-            
-            <!-- Add NCF fields in a clean, organized way -->
-            <xpath expr="//field[@name='payment_reference']" position="after">
-                <group string="📋 Comprobante Fiscal" 
-                       attrs="{'invisible': [('move_type', 'not in', ['out_invoice', 'out_refund'])]}"
-                       class="o_group_col_6">
-                    <field name="use_ncf" 
-                           string="Usar NCF"
-                           widget="boolean_toggle"
-                           help="Activar para usar Número de Comprobante Fiscal"/>
-                    <field name="ncf_fiscal_type_id" 
-                           string="Tipo de Comprobante"
-                           attrs="{'invisible': [('use_ncf', '=', False)], 'required': [('use_ncf', '=', True), ('move_type', '=', 'out_invoice')]}"
-                           options="{'no_create': False, 'no_edit': True}"
-                           placeholder="Seleccione el tipo..."
-                           help="Tipo de comprobante fiscal según DGII"/>
-                    <field name="ncf" 
-                           string="Número NCF"
-                           readonly="1"
-                           attrs="{'invisible': [('ncf', '=', False)]}"
-                           help="Número generado automáticamente"/>
-                </group>
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Simplified tree view -->
-    <record id="view_move_tree_inherit_ncf" model="ir.ui.view">
-        <field name="name">account.move.tree.inherit.ncf</field>
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account.view_move_tree"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='name']" position="after">
-                <field name="ncf" 
-                       string="NCF"
-                       optional="show"/>
-                <field name="ncf_fiscal_type_id" 
-                       string="Tipo NCF"
-                       optional="hide"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
