+++ models/__init__.py
@@ -1,4 +1,6 @@
-from . import ncf_sequence
+# -*- coding: utf-8 -*-
+
 from . import account_move
-from . import res_partner
+from . import ncf_fiscal_type
+from . import ncf_sequence
 from . import res_config_settings
