--- security/ncf_security.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
-        <!-- Definición del modelo wizard para NCF -->
-        <record id="model_ncf_sequence_wizard" model="ir.model">
-            <field name="name">NCF Sequence Wizard</field>
-            <field name="model">ncf.sequence.wizard</field>
-            <field name="state">base</field>
-            <field name="info">Modelo Transient para el Wizard de selección de NCF</field>
-            <field name="modules" eval="['ncf']"/>
-        </record>
-        
-        <!-- Regla de acceso multi-compañía para NCF Sequences -->
-        <record id="ncf_sequence_company_rule" model="ir.rule">
-            <field name="name">NCF Sequence Multi-Company Rule</field>
-            <field name="model_id" ref="ncf.model_ncf_sequence"/>
-            <field name="domain_force">[('company_id', 'in', company_ids)]</field>
-            <field name="groups" eval="[(4, ref('base.group_multi_company'))]"/>
-            <field name="perm_read" eval="True"/>
-            <field name="perm_write" eval="True"/>
-            <field name="perm_create" eval="True"/>
-            <field name="perm_unlink" eval="True"/>
-        </record>
-        
-        <!-- Regla para que los usuarios solo vean secuencias activas -->
-        <record id="ncf_sequence_active_rule" model="ir.rule">
-            <field name="name">NCF Sequence Active Rule</field>
-            <field name="model_id" ref="ncf.model_ncf_sequence"/>
-            <field name="domain_force">[('active', '=', True)]</field>
-            <field name="groups" eval="[(4, ref('account.group_account_invoice'))]"/>
-            <field name="perm_read" eval="True"/>
-            <field name="perm_write" eval="False"/>
-            <field name="perm_create" eval="False"/>
-            <field name="perm_unlink" eval="False"/>
-        </record>
-        
-        <!-- Grupo personalizado para administradores de NCF -->
-        <record id="group_ncf_admin" model="res.groups">
-            <field name="name">NCF Administrator</field>
-            <field name="category_id" ref="base.module_category_accounting"/>
-            <field name="comment">Usuarios con permisos completos sobre la configuración de NCF</field>
-            <field name="implied_ids" eval="[(4, ref('account.group_account_manager'))]"/>
-        </record>
-    </data>
-</odoo>
