+++ views/ncf_sequence_views.xml
@@ -1,66 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <data>
-        <!-- Vista Form de la secuencia NCF -->
-        <record id="view_ncf_sequence_form" model="ir.ui.view">
-            <field name="name">ncf.sequence.form</field>
-            <field name="model">ncf.sequence</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <div class="oe_title">
-                            <label for="name" class="oe_edit_only"/>
-                            <h1><field name="name"/></h1>
-                        </div>
-                        <group>
-                            <group>
-                                <field name="ncf_type"/>
-                                <field name="prefix"/>
-                                <field name="padding"/>
-                                <field name="number_next_actual"/>
-                            </group>
-                            <group>
-                                <field name="company_id" groups="base.group_multi_company"/>
-                                <field name="active"/>
-                            </group>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
+    <record id="view_ncf_sequence_form" model="ir.ui.view">
+        <field name="name">ncf.sequence.form</field>
+        <field name="model">ncf.sequence</field>
+        <field name="arch" type="xml">
+            <form>
+                <sheet>
+                    <group>
+                        <field name="name"/>
+                        <field name="prefix"/>
+                        <field name="padding"/>
+                        <field name="number_next_actual"/>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
 
-        <!-- Vista List (anteriormente Tree) -->
-        <record id="view_ncf_sequence_tree" model="ir.ui.view">
-            <field name="name">ncf.sequence.list</field>
-            <field name="model">ncf.sequence</field>
-            <field name="arch" type="xml">
-                <list>
-                    <field name="name"/>
-                    <field name="prefix"/>
-                    <field name="number_next_actual"/>
-                    <field name="company_id" groups="base.group_multi_company"/>
-                    <field name="active"/>
-                </list>
-            </field>
-        </record>
+    <record id="action_ncf_sequence" model="ir.actions.act_window">
+        <field name="name">Secuencias NCF</field>
+        <field name="res_model">ncf.sequence</field>
+        <field name="view_mode">tree,form</field>
+    </record>
 
-        <!-- Acción ventana -->
-        <record id="action_ncf_sequence" model="ir.actions.act_window">
-            <field name="name">Secuencias NCF</field>
-            <field name="res_model">ncf.sequence</field>
-            <field name="view_mode">list,form</field>
-            <field name="help" type="html">
-                <p class="o_view_nocontent_smiling_face">
-                    Cree su primera secuencia NCF
-                </p>
-            </field>
-        </record>
-
-        <!-- Menú (ajusta el parent si difiere en tu instalación) -->
-        <menuitem id="menu_ncf_sequence" 
-                  name="Secuencias NCF" 
-                  parent="account.menu_finance_configuration" 
-                  action="action_ncf_sequence"
-                  sequence="10"/>
-    </data>
+    <menuitem id="menu_ncf_sequence" name="Secuencias NCF" parent="account.menu_finance_configuration" action="action_ncf_sequence"/>
 </odoo>
