--- views/res_partner_views.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <!-- Herencia de vista de formulario de partner para NCF -->
-        <record id="view_partner_form_inherit_ncf" model="ir.ui.view">
-            <field name="name">res.partner.form.inherit.ncf</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//page[@name='sales_purchases']//group[@name='sale']" position="inside">
-                    <field name="default_ncf_sequence_id" 
-                           string="Secuencia NCF por Defecto"
-                           domain="[('active', '=', True), ('company_id', '=', company_id)]"
-                           context="{'default_company_id': company_id}"
-                           options="{'no_create': True, 'no_edit': True}"
-                           help="Secuencia NCF que se utilizará por defecto para las facturas de este cliente"/>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</odoo>
