--- wizards/__init__.py
@@ -1,1 +1,0 @@
-from . import ncf_sequence_wizard
